requirements.txt
asgiref==3.5.2
boto3==1.24.68
botocore==1.27.68
dj-database-url==0.5.0
Django==3.2
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.13.1
gunicorn==20.1.0
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python3-openid==3.2.0
pytz==2022.2.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sqlparse==0.4.2
stripe==4.1.0
